fixed so error cannod remove rate file does not print to command line

+++ pyMARS/pyMARS.py
@@ -98,8 +98,9 @@ def readin(args='none', **argv):
 
     else:
         print("\n\nFile type not supported")
-    try:
+    if os.path.exists('mass_fractions.hdf5'):
         os.system('rm mass_fractions.hdf5')
+    if os.path.exists('production_rates.hdf5'):
         os.system('rm production_rates.hdf5')
     except Exception:
         pass
